Change to use exceptions from ldap module.
Previously, fakeldap was using exceptions that it defined internally. 

When caller code substitutes the fakeldap for the real one in a test, 
it expects (and tries to catch) the authentic python-ldap exceptions,
not the duplicate ones defined in fakeldap.

Therefore, we've changed fakeldap to throw the authentic exceptions 
as defined in python-ldap.