pypi: Fix ModuleNotFoundError when installing from source dist.
The way we store the package's version number leads to a
ModuleNotFoundError exception when one tries to install the package
from source distributions. Since the version number is only used in
setup.py, we can just store it in setup.py directly, instead of
having to rely on more elaborate schemes.

Thanks to Anders Kaseorg for pointing me to the documentation that
allowed me to fix this! See the following link for details:

https://packaging.python.org/en/latest/guides/single-sourcing-package-version/