fakeldap.py, tests.py: implement anded conditional searching
Re-write `_simple_onelevel_search` into `_multiple_attrs_onelevel_search`
which takes a list of tuples aka a list of conditions that all have to be
met in order for the element to be added to the result list.

Add a second regular expression to `_search_s` that allows multiple
anded conditions to be processed: (&(attr1=val1)(attr2=val2)...)

Implement a short test for the new functionality.